Remove Bindable interface
