+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core/props/data_query/mod.rs
@@ -1,6 +1,5 @@
 //! Data queries are used for requesting props, often for the computation of other props.
 
-mod bindable;
 mod content_filter;
 #[allow(clippy::module_inception)]
 mod data_query;
@@ -8,7 +7,6 @@ mod filters;
 
 use super::*;
 
-pub use bindable::*;
 pub use content_filter::*;
 pub use data_query::*;
 pub use filters::*;
